add schemas.ts
