+++ packages/nextra-theme-docs/src/schemas.ts
@@ -1,0 +1,146 @@
+import type { FC, ReactNode } from 'react';
+import { isValidElement } from 'react'
+import { z } from 'zod'
+import type { NavBarProps } from './components/navbar'
+import type { TOCProps } from './components/toc'
+import type { NextSeoProps } from 'next-seo'
+
+function isReactNode(value: unknown): boolean {
+  return (
+    value == null ||
+    typeof value === 'string' ||
+    isFunction(value) ||
+    isValidElement(value as any)
+  )
+}
+
+function isFunction(value: unknown): boolean {
+  return typeof value === 'function'
+}
+
+const i18nSchema = z.array(
+  z.strictObject({
+    direction: z.enum(['ltr', 'rtl']).optional(),
+    locale: z.string(),
+    name: z.string()
+  })
+)
+
+const reactNode = [
+  isReactNode,
+  { message: 'Must be React.ReactNode or React.FC' }
+] as const
+const fc = [isFunction, { message: 'Must be React.FC' }] as const
+
+export const themeSchema = z.strictObject({
+  banner: z.strictObject({
+    content: z.custom<ReactNode | FC>(...reactNode).optional(),
+    dismissible: z.boolean(),
+    key: z.string()
+  }),
+  chat: z.strictObject({
+    icon: z.custom<ReactNode | FC>(...reactNode),
+    link: z.string().startsWith('https://').optional()
+  }),
+  components: z.record(z.custom<FC>(...fc)).optional(),
+  darkMode: z.boolean(),
+  direction: z.enum(['ltr', 'rtl']),
+  docsRepositoryBase: z.string().startsWith('https://'),
+  editLink: z.strictObject({
+    component: z.custom<
+      FC<{
+        children: ReactNode
+        className?: string
+        filePath?: string
+      }>
+    >(...fc),
+    content: z.custom<ReactNode | FC>(...reactNode)
+  }),
+  faviconGlyph: z.string().optional(),
+  feedback: z.strictObject({
+    content: z.custom<ReactNode | FC>(...reactNode),
+    labels: z.string(),
+    useLink: z.function().returns(z.string())
+  }),
+  footer: z.strictObject({
+    component: z.custom<ReactNode | FC<{ menu: boolean }>>(...reactNode),
+    content: z.custom<ReactNode | FC>(...reactNode)
+  }),
+  gitTimestamp: z.custom<ReactNode | FC<{ timestamp: Date }>>(...reactNode),
+  head: z.custom<ReactNode | FC>(...reactNode),
+  i18n: i18nSchema,
+  logo: z.custom<ReactNode | FC>(...reactNode),
+  logoLink: z.boolean().or(z.string()),
+  main: z.custom<FC<{ children: ReactNode }>>(...fc).optional(),
+  navbar: z.strictObject({
+    component: z.custom<ReactNode | FC<NavBarProps>>(...reactNode),
+    extraContent: z.custom<ReactNode | FC>(...reactNode).optional()
+  }),
+  navigation: z.boolean().or(
+    z.strictObject({
+      next: z.boolean(),
+      prev: z.boolean()
+    })
+  ),
+  nextThemes: z.strictObject({
+    defaultTheme: z.string(),
+    forcedTheme: z.string().optional(),
+    storageKey: z.string()
+  }),
+  notFound: z.strictObject({
+    content: z.custom<ReactNode | FC>(...reactNode),
+    labels: z.string()
+  }),
+  primaryHue: z.number().or(
+    z.strictObject({
+      dark: z.number(),
+      light: z.number()
+    })
+  ),
+  project: z.strictObject({
+    icon: z.custom<ReactNode | FC>(...reactNode),
+    link: z.string().startsWith('https://').optional()
+  }),
+  search: z.strictObject({
+    component: z.custom<ReactNode | FC<{ className?: string }>>(...reactNode),
+    emptyResult: z.custom<ReactNode | FC>(...reactNode),
+    error: z.string().or(z.function().returns(z.string())),
+    loading: z.custom<ReactNode | FC>(...reactNode),
+    // Can't be React component
+    placeholder: z.string().or(z.function().returns(z.string()))
+  }),
+  serverSideError: z.strictObject({
+    content: z.custom<ReactNode | FC>(...reactNode),
+    labels: z.string()
+  }),
+  sidebar: z.strictObject({
+    autoCollapse: z.boolean().optional(),
+    defaultMenuCollapseLevel: z.number().min(1).int(),
+    titleComponent: z.custom<
+      ReactNode | FC<{ title: string; type: string; route: string }>
+    >(...reactNode),
+    toggleButton: z.boolean()
+  }),
+  themeSwitch: z.strictObject({
+    component: z.custom<ReactNode | FC<{ lite?: boolean; className?: string }>>(
+      ...reactNode
+    ),
+    useOptions: themeOptionsSchema.or(z.function().returns(themeOptionsSchema))
+  }),
+  toc: z.strictObject({
+    backToTop: z.boolean(),
+    component: z.custom<ReactNode | FC<TOCProps>>(...reactNode),
+    extraContent: z.custom<ReactNode | FC>(...reactNode),
+    float: z.boolean(),
+    headingComponent: z
+      .custom<FC<{ id: string; children: string }>>(...fc)
+      .optional(),
+    title: z.custom<ReactNode | FC>(...reactNode)
+  }),
+  useNextSeoProps: z.custom<() => NextSeoProps | void>(isFunction)
+})
+
+const publicThemeSchema = themeSchema.deepPartial().extend({
+  // to have `locale` and `text` as required properties
+  i18n: i18nSchema.optional()
+})
